Angled undocking
This adds the ability to drive backwards in a straight line or curve
after undocking, optionally with a random angle.

This can be used to reduce grass wear from undocking and also to get to
an area with better GPS reception.

This introduces 4 new environment variables for configuration:
- OM_UNDOCK_ANGLED_DISTANCE:
  The additional second stage distance to drive at an angle for
  undocking. This needs to be large enough for the robot to have GPS
  reception. This section may still be driven without GPS so don't
  expect high positional accuracy.
- OM_UNDOCK_ANGLE:
  The angle at which to drive for the additional distance (neg values
  are to the left of the dock, pos to the right).
- OM_UNDOCK_FIXED_ANGLE:
  If false will vary the undocking angle between +abs(OM_UNDOCK_ANGLE)
  to -abs(OM_UNDOCK_ANGLE) on subsequent undocks to reduce grass wear.
- OM_UNDOCK_USE_CURVE:
  If true will use a curved second stage undock move, if false will use
  a straight undock move.

This also restores one more config variable that already existed but was
erroneously removed in a previous commit: OM_DOCKING_APPROACH_DISTANCE